requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.2
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.6
requests==2.31.0
llama-parse==0.3.0
chromadb==0.4.15
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.36.0
numpy==1.24.3
pandas==2.0.3